% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilterCarScore.R
\docType{data}
\name{FilterCarScore}
\alias{FilterCarScore}
\alias{mlr_filters_carscore}
\title{Conditional Mutual Information Based Feature Selection Filter}
\format{\link[R6:R6Class]{R6::R6Class} inheriting from \link{Filter}.}
\usage{
FilterCarScore
}
\description{
Calculates the Correlation-Adjusted (marginal) coRelation scores (short CAR scores)
implemented in \code{\link[care:carscore]{care::carscore()}} in package \CRANpkg{care}.
The CAR scores for a set of features are defined as the correlations between the target
and the decorrelated features. The filter returns the absolute value of the calculated scores.

Argument \code{verbose} defaults to \code{FALSE}.
}
\examples{
task = mlr3::mlr_tasks$get("mtcars")
filter = FilterCarScore$new()
filter$calculate(task)
head(as.data.table(filter), 3)

## changing filter settings
filter = FilterCarScore$new()
filter$param_set$values = list("diagonal" = TRUE)
filter$calculate(task)
head(as.data.table(filter), 3)
}
\seealso{
Other Filter: \code{\link{FilterAUC}},
  \code{\link{FilterAnova}}, \code{\link{FilterCMIM}},
  \code{\link{FilterCorrelation}},
  \code{\link{FilterDISR}}, \code{\link{FilterImportance}},
  \code{\link{FilterInformationGain}},
  \code{\link{FilterJMIM}}, \code{\link{FilterJMI}},
  \code{\link{FilterKruskalTest}}, \code{\link{FilterMIM}},
  \code{\link{FilterMRMR}}, \code{\link{FilterNJMIM}},
  \code{\link{FilterPerformance}},
  \code{\link{FilterVariance}}, \code{\link{Filter}},
  \code{\link{mlr_filters}}
}
\concept{Filter}
\keyword{datasets}
