% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilterGainRatio.R
\docType{data}
\name{FilterGainRatio}
\alias{FilterGainRatio}
\alias{mlr_filters_gain_ratio}
\title{Gain Ratio Filter}
\format{\link[R6:R6Class]{R6::R6Class} inheriting from \link{Filter}.}
\usage{
FilterGainRatio
}
\description{
Gain Ratio filter calling \code{\link[FSelectorRcpp:information_gain]{FSelectorRcpp::information_gain()}} from package \CRANpkg{FSelectorRcpp}.
Argument \code{equal} defaults to \code{FALSE} for classification tasks, and to \code{TRUE} for regression tasks.
}
\examples{
task = mlr3::mlr_tasks$get("pima")
filter = FilterGainRatio$new()
filter$calculate(task)
head(filter$scores, 3)
as.data.table(filter)
}
\seealso{
Other Filter: \code{\link{FilterAUC}},
  \code{\link{FilterAnova}}, \code{\link{FilterCMIM}},
  \code{\link{FilterCarScore}},
  \code{\link{FilterCorrelation}},
  \code{\link{FilterDISR}}, \code{\link{FilterImportance}},
  \code{\link{FilterInformationGain}},
  \code{\link{FilterJMIM}}, \code{\link{FilterJMI}},
  \code{\link{FilterKruskalTest}}, \code{\link{FilterMIM}},
  \code{\link{FilterMRMR}}, \code{\link{FilterNJMIM}},
  \code{\link{FilterPerformance}},
  \code{\link{FilterSymmetricalUncertainty}},
  \code{\link{FilterVariance}}, \code{\link{Filter}},
  \code{\link{mlr_filters}}
}
\concept{Filter}
\keyword{datasets}
