% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilterJMIM.R
\docType{data}
\name{FilterJMIM}
\alias{FilterJMIM}
\alias{mlr_filters_jmim}
\title{Minimal Joint Mutual Information Maximisation Filter}
\format{\link[R6:R6Class]{R6::R6Class} inheriting from \link{Filter}.}
\usage{
FilterJMIM
}
\description{
Minimal joint mutual information maximisation filter calling \code{\link[praznik:JMIM]{praznik::JMIM()}} in package \CRANpkg{praznik}.

This filter supports partial scoring (see \link{Filter}).
}
\examples{
task = mlr3::mlr_tasks$get("iris")
filter = FilterJMIM$new()
filter$calculate(task, nfeat = 2)
head(filter$scores, 2)
as.data.table(filter)
}
\seealso{
Other Filter: \code{\link{FilterAUC}},
  \code{\link{FilterAnova}}, \code{\link{FilterCMIM}},
  \code{\link{FilterCarScore}},
  \code{\link{FilterCorrelation}},
  \code{\link{FilterDISR}}, \code{\link{FilterGainRatio}},
  \code{\link{FilterImportance}},
  \code{\link{FilterInformationGain}},
  \code{\link{FilterJMI}}, \code{\link{FilterKruskalTest}},
  \code{\link{FilterMIM}}, \code{\link{FilterMRMR}},
  \code{\link{FilterNJMIM}},
  \code{\link{FilterPerformance}},
  \code{\link{FilterSymmetricalUncertainty}},
  \code{\link{FilterVariance}}, \code{\link{Filter}},
  \code{\link{mlr_filters}}
}
\concept{Filter}
\keyword{datasets}
