% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilterNJMIM.R
\docType{data}
\name{FilterNJMIM}
\alias{FilterNJMIM}
\alias{mlr_filters_njmim}
\title{Minimal Normalised Joint Mutual Information Maximisation Filter}
\format{\link[R6:R6Class]{R6::R6Class} inheriting from \link{Filter}.}
\usage{
FilterNJMIM
}
\description{
Minimal normalised joint mutual information maximisation filter.
Calls \code{\link[praznik:NJMIM]{praznik::NJMIM()}}.
}
\details{
This filter supports partial scoring via hyperparameter \code{k}. For
internal reasons, \code{k} is not exposed in the ParamSet. Instead, the generic
hyperparameter \code{nfeat} will be populated to the filter for partial scoring
calculation. By default all filter values are calculated (\code{nfeat = 1})
}
\examples{
task = mlr3::mlr_tasks$get("iris")
filter = FilterNJMIM$new()
filter$calculate(task)
as.data.table(filter)[1:3]
}
\seealso{
Other Filter: \code{\link{FilterAUC}},
  \code{\link{FilterCMIM}}, \code{\link{FilterCarScore}},
  \code{\link{FilterCorrelation}},
  \code{\link{FilterDISR}}, \code{\link{FilterEmbedded}},
  \code{\link{FilterGainRatio}},
  \code{\link{FilterInformationGain}},
  \code{\link{FilterJMIM}}, \code{\link{FilterJMI}},
  \code{\link{FilterKruskalTest}}, \code{\link{FilterMIM}},
  \code{\link{FilterMRMR}},
  \code{\link{FilterSymmetricalUncertainty}},
  \code{\link{FilterVariance}}, \code{\link{Filter}},
  \code{\link{mlr_filters}}
}
\concept{Filter}
\keyword{datasets}
