% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flt.R
\name{flt}
\alias{flt}
\alias{flts}
\title{Syntactic Sugar for Filter Construction}
\usage{
flt(.key, ...)

flts(.keys, ...)
}
\value{
\link{Filter}.
}
\description{
These functions complements \link{mlr_filters} with a function in the spirit of \link[mlr3:mlr_sugar]{mlr3::mlr_sugar}.
}
\examples{
flt("correlation", method = "kendall")
flts(c("mrmr", "jmim"))
}
