% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{mlr3filters-package}
\alias{mlr3filters}
\alias{mlr3filters-package}
\title{mlr3filters: Filter based feature selection for 'mlr3'}
\description{
Implements methods for feature selection and
    filtering in 'mlr3'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mlr-org/mlr3filters}
  \item Report bugs at \url{https://github.com/mlr-org/mlr3filters/issues}
}

}
\author{
\strong{Maintainer}: Patrick Schratz \email{patrick.schratz@gmail.com} (0000-0003-0748-6624)

Authors:
\itemize{
  \item Michel Lang \email{michellang@gmail.com} (0000-0001-9754-0393)
  \item Bernd Bischl \email{bernd_bischl@gmx.net} (0000-0001-6002-6980)
}

}
