% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilterAnova.R
\name{mlr_filters_anova}
\alias{mlr_filters_anova}
\alias{FilterAnova}
\title{ANOVA F-Test Filter}
\format{\link[R6:R6Class]{R6::R6Class} inheriting from \link{Filter}.}
\description{
ANOVA F-Test filter calling \code{\link[stats:aov]{stats::aov()}}. Note that this is
equivalent to a \eqn{t}-test for binary classification.

The filter value is \code{-log10(p)} where \code{p} is the \eqn{p}-value. This transformation
is necessary to ensure numerical stability for very small \eqn{p}-values.
}
\section{Construction}{
\preformatted{FilterAnova$new()
mlr_filters$get("anova")
flt("anova")
}
}

\examples{
task = mlr3::tsk("iris")
filter = flt("anova")
filter$calculate(task)
head(as.data.table(filter), 3)

# transform to p-value
10^(-filter$scores)
}
\seealso{
\link[mlr3misc:Dictionary]{Dictionary} of \link[=Filter]{Filters}: \link{mlr_filters}

Other Filter: 
\code{\link{Filter}},
\code{\link{mlr_filters_auc}},
\code{\link{mlr_filters_carscore}},
\code{\link{mlr_filters_cmim}},
\code{\link{mlr_filters_correlation}},
\code{\link{mlr_filters_disr}},
\code{\link{mlr_filters_find_correlation}},
\code{\link{mlr_filters_information_gain}},
\code{\link{mlr_filters_jmim}},
\code{\link{mlr_filters_jmi}},
\code{\link{mlr_filters_kruskal_test}},
\code{\link{mlr_filters_mim}},
\code{\link{mlr_filters_mrmr}},
\code{\link{mlr_filters_njmim}},
\code{\link{mlr_filters_performance}},
\code{\link{mlr_filters_variable_importance}},
\code{\link{mlr_filters_variance}},
\code{\link{mlr_filters}}
}
\concept{Filter}
