% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilterAUC.R
\name{mlr_filters_auc}
\alias{mlr_filters_auc}
\alias{FilterAUC}
\title{AUC Filter}
\format{\link[R6:R6Class]{R6::R6Class} inheriting from \link{Filter}.}
\description{
Area under the (ROC) Curve filter, analogously to \code{\link[mlr3measures:auc]{mlr3measures::auc()}} from \CRANpkg{mlr3measures}.
Missing values of the features are removed before calculating the AUC.
If the AUC is undefined for the input, it is set to 0.5 (random classifier).
The absolute value of the difference between the AUC and 0.5 is used as final filter value.
}
\section{Construction}{
\preformatted{FilterAUC$new()
mlr_filters$get("auc")
flt("auc")
}
}

\examples{
task = mlr3::tsk("pima")
filter = flt("auc")
filter$calculate(task)
head(as.data.table(filter), 3)
}
\seealso{
\link[mlr3misc:Dictionary]{Dictionary} of \link[=Filter]{Filters}: \link{mlr_filters}

Other Filter: 
\code{\link{Filter}},
\code{\link{mlr_filters_anova}},
\code{\link{mlr_filters_carscore}},
\code{\link{mlr_filters_cmim}},
\code{\link{mlr_filters_correlation}},
\code{\link{mlr_filters_disr}},
\code{\link{mlr_filters_information_gain}},
\code{\link{mlr_filters_jmim}},
\code{\link{mlr_filters_jmi}},
\code{\link{mlr_filters_kruskal_test}},
\code{\link{mlr_filters_mim}},
\code{\link{mlr_filters_mrmr}},
\code{\link{mlr_filters_njmim}},
\code{\link{mlr_filters_performance}},
\code{\link{mlr_filters_variable_importance}},
\code{\link{mlr_filters_variance}},
\code{\link{mlr_filters}}
}
\concept{Filter}
