% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilterCorrelation.R, R/FilterFindCorrelation.R
\name{mlr_filters_correlation}
\alias{mlr_filters_correlation}
\alias{FilterCorrelation}
\alias{FilterFindCorrelation}
\title{Correlation Filter}
\format{\link[R6:R6Class]{R6::R6Class} inheriting from \link{Filter}.}
\description{
Simple correlation filter calling \code{\link[stats:cor]{stats::cor()}}.
The filter score is the absolute value of the correlation.

Simple filter emulating \code{caret::findCorrelation(exact = FALSE_)}.

This gives each feature a score that is the \emph{negative} of the cutoff value
for which it is excluded when using \code{caret::findCorrelation()}. The negative
is used because \code{caret::findCorrelation()} excludes everything \emph{above} a cutoff,
while filters exclude everything below a cutoff. Therefore, \code{caret::findCorrelation(cutoff = 0.9)}
lists the same features that are excluded with \code{FilterFindCorrelation} at score -0.9.
}
\section{Construction}{
\preformatted{FilterCorrelation$new()
mlr_filters$get("correlation")
flt("correlation")
}

\preformatted{FilterFindCorrelation$new()
mlr_filters$get("findcorrelation")
flt("findcorrelation")
}
}

\examples{
## Pearson (default)
task = mlr3::tsk("mtcars")
filter = flt("correlation")
filter$calculate(task)
as.data.table(filter)

## Spearman
filter = FilterCorrelation$new()
filter$param_set$values = list("method" = "spearman")
filter$calculate(task)
as.data.table(filter)
## Pearson (default)
task = mlr3::tsk("mtcars")
filter = flt("findcorrelation")
filter$calculate(task)
as.data.table(filter)

## Spearman
filter = FilterFindCorrelation$new()
filter$param_set$values = list("method" = "spearman")
filter$calculate(task)
as.data.table(filter)
}
\seealso{
\link[mlr3misc:Dictionary]{Dictionary} of \link[=Filter]{Filters}: \link{mlr_filters}

\link[mlr3misc:Dictionary]{Dictionary} of \link[=Filter]{Filters}: \link{mlr_filters}

Other Filter: 
\code{\link{Filter}},
\code{\link{mlr_filters_anova}},
\code{\link{mlr_filters_auc}},
\code{\link{mlr_filters_carscore}},
\code{\link{mlr_filters_cmim}},
\code{\link{mlr_filters_disr}},
\code{\link{mlr_filters_information_gain}},
\code{\link{mlr_filters_jmim}},
\code{\link{mlr_filters_jmi}},
\code{\link{mlr_filters_kruskal_test}},
\code{\link{mlr_filters_mim}},
\code{\link{mlr_filters_mrmr}},
\code{\link{mlr_filters_njmim}},
\code{\link{mlr_filters_performance}},
\code{\link{mlr_filters_variable_importance}},
\code{\link{mlr_filters_variance}},
\code{\link{mlr_filters}}

Other Filter: 
\code{\link{Filter}},
\code{\link{mlr_filters_anova}},
\code{\link{mlr_filters_auc}},
\code{\link{mlr_filters_carscore}},
\code{\link{mlr_filters_cmim}},
\code{\link{mlr_filters_disr}},
\code{\link{mlr_filters_information_gain}},
\code{\link{mlr_filters_jmim}},
\code{\link{mlr_filters_jmi}},
\code{\link{mlr_filters_kruskal_test}},
\code{\link{mlr_filters_mim}},
\code{\link{mlr_filters_mrmr}},
\code{\link{mlr_filters_njmim}},
\code{\link{mlr_filters_performance}},
\code{\link{mlr_filters_variable_importance}},
\code{\link{mlr_filters_variance}},
\code{\link{mlr_filters}}
}
\concept{Filter}
