% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilterFindCorrelation.R
\name{mlr_filters_find_correlation}
\alias{mlr_filters_find_correlation}
\alias{FilterFindCorrelation}
\title{Correlation Filter}
\format{\link[R6:R6Class]{R6::R6Class} inheriting from \link{Filter}.}
\description{
Simple filter emulating \code{caret::findCorrelation(exact = FALSE)}.

This gives each feature a score between 0 and 1 that is \emph{one minus} the
cutoff value for which it is excluded when using \code{\link[caret:findCorrelation]{caret::findCorrelation()}}.
The negative is used because \code{\link[caret:findCorrelation]{caret::findCorrelation()}} excludes everything
\emph{above} a cutoff, while filters exclude everything below a cutoff.
Here the filter scores are shifted by +1 to get positive values for to align
with the way other filters work.

Subsequently \code{caret::findCorrelation(cutoff = 0.9)} lists the same features
that are excluded with \code{FilterFindCorrelation} at score 0.1 (= 1 - 0.9).
}
\section{Construction}{
\preformatted{FilterFindCorrelation$new()
mlr_filters$get("findCorrelation")
flt("findCorrelation")
}
}

\examples{
## Pearson (default)
task = mlr3::tsk("mtcars")
filter = flt("findCorrelation")
filter$calculate(task)
as.data.table(filter)

## Spearman
filter = flt("findCorrelation", method = "spearman")
filter$calculate(task)
as.data.table(filter)
}
\seealso{
\link[mlr3misc:Dictionary]{Dictionary} of \link[=Filter]{Filters}: \link{mlr_filters}

Other Filter: 
\code{\link{Filter}},
\code{\link{mlr_filters_anova}},
\code{\link{mlr_filters_auc}},
\code{\link{mlr_filters_carscore}},
\code{\link{mlr_filters_cmim}},
\code{\link{mlr_filters_correlation}},
\code{\link{mlr_filters_disr}},
\code{\link{mlr_filters_information_gain}},
\code{\link{mlr_filters_jmim}},
\code{\link{mlr_filters_jmi}},
\code{\link{mlr_filters_kruskal_test}},
\code{\link{mlr_filters_mim}},
\code{\link{mlr_filters_mrmr}},
\code{\link{mlr_filters_njmim}},
\code{\link{mlr_filters_performance}},
\code{\link{mlr_filters_variable_importance}},
\code{\link{mlr_filters_variance}},
\code{\link{mlr_filters}}
}
\concept{Filter}
