% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilterGausscovF1st.R
\name{mlr_filters_gauscovf1st}
\alias{mlr_filters_gauscovf1st}
\alias{FilterGausscovF1st}
\title{Gaussian Covariate Stepwise Selection}
\description{
Gaussian covariate filter calling
\code{\link[gausscov:f1st]{gausscov::f1st()}} in package \CRANpkg{gausscov}. Please read
original documentation for the function form the \CRANpkg{gausscov}
package.

Filter supports only regression tasks.
}
\examples{
if (requireNamespace("FilterGausscovF1st")) {
  ## Gaussian Covariance (default)
  task = mlr3::tsk("mtcars")
  filter = flt("gausscov_f1st")
  filter$calculate(task)
  head(filter$scores, 3)
  as.data.table(filter)
}
}
\seealso{
\itemize{
\item \link[mlr3pipelines:mlr_pipeops_filter]{PipeOpFilter} for filter-based feature selection.
\item \link[mlr3misc:Dictionary]{Dictionary} of \link[=Filter]{Filters}: \link{mlr_filters}
}

Other Filter: 
\code{\link{Filter}},
\code{\link{mlr_filters_anova}},
\code{\link{mlr_filters_auc}},
\code{\link{mlr_filters_carscore}},
\code{\link{mlr_filters_carsurvscore}},
\code{\link{mlr_filters_cmim}},
\code{\link{mlr_filters_correlation}},
\code{\link{mlr_filters_disr}},
\code{\link{mlr_filters_find_correlation}},
\code{\link{mlr_filters_importance}},
\code{\link{mlr_filters_information_gain}},
\code{\link{mlr_filters_jmim}},
\code{\link{mlr_filters_jmi}},
\code{\link{mlr_filters_kruskal_test}},
\code{\link{mlr_filters_mim}},
\code{\link{mlr_filters_mrmr}},
\code{\link{mlr_filters_njmim}},
\code{\link{mlr_filters_performance}},
\code{\link{mlr_filters_permutation}},
\code{\link{mlr_filters_relief}},
\code{\link{mlr_filters_selected_features}},
\code{\link{mlr_filters_variance}},
\code{\link{mlr_filters}}
}
\concept{Filter}
\section{Super class}{
\code{\link[mlr3filters:Filter]{mlr3filters::Filter}} -> \code{FilterGausscovF1st}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-FilterGausscovF1st-new}{\code{FilterGausscovF1st$new()}}
\item \href{#method-FilterGausscovF1st-clone}{\code{FilterGausscovF1st$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="mlr3filters" data-topic="Filter" data-id="calculate"><a href='../../mlr3filters/html/Filter.html#method-Filter-calculate'><code>mlr3filters::Filter$calculate()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3filters" data-topic="Filter" data-id="format"><a href='../../mlr3filters/html/Filter.html#method-Filter-format'><code>mlr3filters::Filter$format()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3filters" data-topic="Filter" data-id="help"><a href='../../mlr3filters/html/Filter.html#method-Filter-help'><code>mlr3filters::Filter$help()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3filters" data-topic="Filter" data-id="print"><a href='../../mlr3filters/html/Filter.html#method-Filter-print'><code>mlr3filters::Filter$print()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FilterGausscovF1st-new"></a>}}
\if{latex}{\out{\hypertarget{method-FilterGausscovF1st-new}{}}}
\subsection{Method \code{new()}}{
Create a GaussCov object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FilterGausscovF1st$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FilterGausscovF1st-clone"></a>}}
\if{latex}{\out{\hypertarget{method-FilterGausscovF1st-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FilterGausscovF1st$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
