% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilterImportance.R
\name{mlr_filters_variable_importance}
\alias{mlr_filters_variable_importance}
\alias{FilterImportance}
\title{Filter for Embedded Feature Selection via Variable Importance}
\format{\link[R6:R6Class]{R6::R6Class} inheriting from \link{Filter}.}
\description{
Variable Importance filter using embedded feature selection of
machine learning algorithms. Takes a \link[mlr3:Learner]{mlr3::Learner} which is capable of
extracting the variable importance (property "importance"), fits the model
and extracts the importance values to use as filter scores.
}
\section{Construction}{
\preformatted{FilterImportance$new(learner = mlr3::lrn("classif.rpart"))
mlr_filters$get("importance")
flt("importance")
}
\itemize{
\item \code{learner} :: \link[mlr3:Learner]{mlr3::Learner}\cr
Learner to extract the importance values from.
}
}

\examples{
task = mlr3::tsk("iris")
learner = mlr3::lrn("classif.rpart")
filter = flt("importance", learner = learner)
filter$calculate(task)
as.data.table(filter)
}
\seealso{
\link[mlr3misc:Dictionary]{Dictionary} of \link[=Filter]{Filters}: \link{mlr_filters}

Other Filter: 
\code{\link{Filter}},
\code{\link{mlr_filters_anova}},
\code{\link{mlr_filters_auc}},
\code{\link{mlr_filters_carscore}},
\code{\link{mlr_filters_cmim}},
\code{\link{mlr_filters_correlation}},
\code{\link{mlr_filters_disr}},
\code{\link{mlr_filters_information_gain}},
\code{\link{mlr_filters_jmim}},
\code{\link{mlr_filters_jmi}},
\code{\link{mlr_filters_kruskal_test}},
\code{\link{mlr_filters_mim}},
\code{\link{mlr_filters_mrmr}},
\code{\link{mlr_filters_njmim}},
\code{\link{mlr_filters_performance}},
\code{\link{mlr_filters_variance}},
\code{\link{mlr_filters}}
}
\concept{Filter}
