% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Filter.R
\docType{data}
\name{Filter}
\alias{Filter}
\title{Filter Base Class}
\format{\link[R6:R6Class]{R6::R6Class} object.}
\description{
Base class for filters. Predefined filters are stored in
\link{mlr_filters}. Filters calculate a score for each feature of a dataset and
return a ranked scoring. By default filter values for all features will be
calculated unless set differently during Filter construction.
}
\section{Construction}{
\preformatted{f = Filter$new(id, task_type, param_set, param_vals, feature_types, packages)
}
\itemize{
\item \code{id} :: \code{character(1)}\cr Identifier for the filter.
\item \code{task_type} :: \code{character(1)}\cr Type of the task the filter can operator
on. E.g., \code{"classif"} or \code{"regr"}.
\item \code{param_set} :: \link[paradox:ParamSet]{paradox::ParamSet}\cr Set of hyperparameters.
\item \code{param_vals} :: named \code{list()}\cr Named list of hyperparameter settings.
\item \code{feature_types} :: \code{character()}\cr Feature types the filter operates on.
Must be a subset of
\code{\link[mlr3:mlr_reflections]{mlr_reflections$task_feature_types}}.
\item \code{task_properties} :: \code{character()}\cr Required task properties, see
\link[mlr3:Task]{mlr3::Task}. Must be a subset of
\code{\link[mlr3:mlr_reflections]{mlr_reflections$task_properties}}.
\item \code{packages} :: \code{character()}\cr Set of required packages. Note that these
packages will be loaded via \code{\link[=requireNamespace]{requireNamespace()}}, and are not attached.
}
}

\section{Fields}{


All fields from the Constructor and
\itemize{
\item \code{scores} :: \link[data.table:data.table]{data.table::data.table}\cr Stores the calculated filter
score values. The scores are sorted in decreasing order, with tied values
in a random order.
}
}

\section{Methods}{

\itemize{
\item \code{calculate(task, nfeat = NULL)}\cr (\link{Task}, \code{numeric(1)}) -> \link{Filter}\cr
Calculates the filter score values for the provided \link{Task} and stores them
in field \code{scores}. Some filter support partial scoring via argument \code{n}.
}
}

\seealso{
Other Filter: \code{\link{FilterAUC}},
  \code{\link{FilterCMIM}}, \code{\link{FilterCarScore}},
  \code{\link{FilterCorrelation}},
  \code{\link{FilterDISR}}, \code{\link{FilterEmbedded}},
  \code{\link{FilterGainRatio}},
  \code{\link{FilterInformationGain}},
  \code{\link{FilterJMIM}}, \code{\link{FilterJMI}},
  \code{\link{FilterKruskalTest}}, \code{\link{FilterMIM}},
  \code{\link{FilterMRMR}}, \code{\link{FilterNJMIM}},
  \code{\link{FilterSymmetricalUncertainty}},
  \code{\link{FilterVariance}}, \code{\link{mlr_filters}}
}
\concept{Filter}
\keyword{datasets}
