% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilterAUC.R
\docType{data}
\name{FilterAUC}
\alias{FilterAUC}
\alias{mlr_filters_auc}
\title{AUC Filter}
\format{\link[R6:R6Class]{R6::R6Class} inheriting from \link{Filter}.}
\usage{
FilterAUC
}
\description{
Area under the (ROC) Curve filter calling \code{\link[Metrics:auc]{Metrics::auc()}} from package \CRANpkg{Metrics}.
Returns the absolute value of the difference between the AUC and 0.5.
}
\examples{
task = mlr3::mlr_tasks$get("pima")
filter = FilterAUC$new()
filter$calculate(task)
head(as.data.table(filter), 3)
}
\seealso{
Other Filter: \code{\link{FilterAnova}},
  \code{\link{FilterCMIM}}, \code{\link{FilterCarScore}},
  \code{\link{FilterCorrelation}},
  \code{\link{FilterDISR}}, \code{\link{FilterImportance}},
  \code{\link{FilterInformationGain}},
  \code{\link{FilterJMIM}}, \code{\link{FilterJMI}},
  \code{\link{FilterKruskalTest}}, \code{\link{FilterMIM}},
  \code{\link{FilterMRMR}}, \code{\link{FilterNJMIM}},
  \code{\link{FilterPerformance}},
  \code{\link{FilterVariance}}, \code{\link{Filter}},
  \code{\link{mlr_filters}}
}
\concept{Filter}
\keyword{datasets}
