% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilterCMIM.R
\docType{data}
\name{FilterCMIM}
\alias{FilterCMIM}
\alias{mlr_filters_cmim}
\title{Minimal Conditional Mutual Information Filter}
\format{\link[R6:R6Class]{R6::R6Class} inheriting from \link{Filter}.}
\usage{
FilterCMIM
}
\description{
Minimal conditional mutual information maximisation filter calling \code{\link[praznik:CMIM]{praznik::CMIM()}} from package \CRANpkg{praznik}.

This filter supports partial scoring (see \link{Filter}).
}
\examples{
task = mlr3::mlr_tasks$get("iris")
filter = FilterCMIM$new()
filter$calculate(task, nfeat = 2)

head(filter$scores, 2)
as.data.table(filter)
}
\seealso{
Other Filter: \code{\link{FilterAUC}},
  \code{\link{FilterCarScore}},
  \code{\link{FilterCorrelation}},
  \code{\link{FilterDISR}}, \code{\link{FilterEmbedded}},
  \code{\link{FilterGainRatio}},
  \code{\link{FilterInformationGain}},
  \code{\link{FilterJMIM}}, \code{\link{FilterJMI}},
  \code{\link{FilterKruskalTest}}, \code{\link{FilterMIM}},
  \code{\link{FilterMRMR}}, \code{\link{FilterNJMIM}},
  \code{\link{FilterSymmetricalUncertainty}},
  \code{\link{FilterVariance}}, \code{\link{Filter}},
  \code{\link{mlr_filters}}
}
\concept{Filter}
\keyword{datasets}
