% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilterJMI.R
\docType{data}
\name{FilterJMI}
\alias{FilterJMI}
\alias{mlr_filters_jmi}
\title{Joint Mutual Information Filter}
\format{\link[R6:R6Class]{R6::R6Class} inheriting from \link{Filter}.}
\usage{
FilterJMI
}
\description{
Joint mutual information filter calling \code{\link[praznik:JMI]{praznik::JMI()}} in package \CRANpkg{praznik}.

This filter supports partial scoring (see \link{Filter}).
}
\examples{
task = mlr3::mlr_tasks$get("iris")
filter = FilterJMI$new()
filter$calculate(task, nfeat = 2)
as.data.table(filter)
}
\seealso{
Other Filter: \code{\link{FilterAUC}},
  \code{\link{FilterAnova}}, \code{\link{FilterCMIM}},
  \code{\link{FilterCarScore}},
  \code{\link{FilterCorrelation}},
  \code{\link{FilterDISR}}, \code{\link{FilterImportance}},
  \code{\link{FilterInformationGain}},
  \code{\link{FilterJMIM}},
  \code{\link{FilterKruskalTest}}, \code{\link{FilterMIM}},
  \code{\link{FilterMRMR}}, \code{\link{FilterNJMIM}},
  \code{\link{FilterPerformance}},
  \code{\link{FilterVariance}}, \code{\link{Filter}},
  \code{\link{mlr_filters}}
}
\concept{Filter}
\keyword{datasets}
