% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilterPerformance.R
\docType{data}
\name{FilterPerformance}
\alias{FilterPerformance}
\alias{mlr_filters_performance}
\title{Predictive Performance Filter}
\format{\link[R6:R6Class]{R6::R6Class} inheriting from \link{Filter}.}
\usage{
FilterPerformance
}
\description{
Filter which uses the predictive performance of a
\link[mlr3:Learner]{mlr3::Learner} as filter score. Performs a \code{\link[mlr3:resample]{mlr3::resample()}} for each
feature separately. The filter score is the aggregated performance of the
\link[mlr3:Measure]{mlr3::Measure}, or the negated aggregated performance if the measure which
are to be minimized.
}
\examples{
task = mlr3::mlr_tasks$get("iris")
learner = mlr3::mlr_learners$get("classif.rpart")
filter = FilterPerformance$new(learner = learner)
filter$calculate(task)
as.data.table(filter)
}
\seealso{
Other Filter: \code{\link{FilterAUC}},
  \code{\link{FilterAnova}}, \code{\link{FilterCMIM}},
  \code{\link{FilterCarScore}},
  \code{\link{FilterCorrelation}},
  \code{\link{FilterDISR}}, \code{\link{FilterImportance}},
  \code{\link{FilterInformationGain}},
  \code{\link{FilterJMIM}}, \code{\link{FilterJMI}},
  \code{\link{FilterKruskalTest}}, \code{\link{FilterMIM}},
  \code{\link{FilterMRMR}}, \code{\link{FilterNJMIM}},
  \code{\link{FilterVariance}}, \code{\link{Filter}},
  \code{\link{mlr_filters}}
}
\concept{Filter}
\keyword{datasets}
