% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flt.R
\name{flt}
\alias{flt}
\title{Syntactic Sugar for Filter Construction}
\usage{
flt(.key, ...)
}
\arguments{
\item{.key}{:: \code{character(1)}\cr
Key passed to the respective \link[mlr3misc:Dictionary]{mlr3misc::Dictionary} to retrieve the object.}

\item{...}{:: named \code{list()}\cr
Named arguments passed to the constructor, to be set as parameters in the \link[paradox:ParamSet]{paradox::ParamSet}, or to be set as public field.
See \code{\link[mlr3misc:dictionary_sugar]{mlr3misc::dictionary_sugar()}} for more details.}
}
\value{
\link{Filter}.
}
\description{
This function complements \link{mlr_filters} with a function in the spirit of \link[mlr3:mlr_sugar]{mlr3::mlr_sugar}.
}
\examples{
flt("correlation", method = "kendall")
}
