% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{mlr3filters-package}
\alias{mlr3filters}
\alias{mlr3filters-package}
\title{mlr3filters: Filter Based Feature Selection for 'mlr3'}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Extends 'mlr3' with filter methods for feature selection. Besides standalone filter methods built-in methods of any machine-learning algorithm are supported. Partial scoring of multivariate filter methods is supported.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://mlr3filters.mlr-org.com}
  \item \url{https://github.com/mlr-org/mlr3filters}
  \item Report bugs at \url{https://github.com/mlr-org/mlr3filters/issues}
}

}
\author{
\strong{Maintainer}: Michel Lang \email{michellang@gmail.com} (\href{https://orcid.org/0000-0001-9754-0393}{ORCID})

Authors:
\itemize{
  \item Patrick Schratz \email{patrick.schratz@gmail.com} (\href{https://orcid.org/0000-0003-0748-6624}{ORCID})
  \item Bernd Bischl \email{bernd_bischl@gmx.net} (\href{https://orcid.org/0000-0001-6002-6980}{ORCID})
  \item Martin Binder \email{mlr.developer@mb706.com}
}

}
