% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlr_filters.R
\docType{data}
\name{mlr_filters}
\alias{mlr_filters}
\title{Dictionary of Filters}
\format{\link{R6Class} object}
\usage{
mlr_filters
}
\description{
A simple \link{Dictionary} storing objects of class \link{Filter}.
Each Filter has an associated help page, see \code{mlr_filters_[id]}.

This dictionary can get populated with additional filters by add-on packages.

For a more convenient way to retrieve and construct filters, see \code{\link[=flt]{flt()}}.
}
\section{Usage}{


See \link{Dictionary}.
}

\examples{
mlr_filters$keys()
as.data.table(mlr_filters)
mlr_filters$get("mim")
flt("anova")
}
\seealso{
Other Filter: \code{\link{FilterAUC}},
  \code{\link{FilterAnova}}, \code{\link{FilterCMIM}},
  \code{\link{FilterCarScore}},
  \code{\link{FilterCorrelation}},
  \code{\link{FilterDISR}}, \code{\link{FilterImportance}},
  \code{\link{FilterInformationGain}},
  \code{\link{FilterJMIM}}, \code{\link{FilterJMI}},
  \code{\link{FilterKruskalTest}}, \code{\link{FilterMIM}},
  \code{\link{FilterMRMR}}, \code{\link{FilterNJMIM}},
  \code{\link{FilterPerformance}},
  \code{\link{FilterVariance}}, \code{\link{Filter}}
}
\concept{Dictionary}
\concept{Filter}
\keyword{datasets}
